Define configuration as an interface
Storing configuration in a file is one way to do it.  It's easy to
understand and easy to get started.  Unfortunately, it's not a very
flexible way to configure the component.  Some instances may want to
store config in a database or generate config by running a script.  I
don't want the component depend to heavily on any one implementation.