Choose SMS provider without message participants
I originally thought that users would want the ability to choose an
SMS provider based on the participants in a message.  That might still
be desirable but the API doesn't work for incoming messages.  That's
because the provider is the only one that can parse an incoming HTTP
request to determine who the participants are.

In practice, nearly all users will have a single SMS provider so this
shouldn't be a major limitation.