Let it Crash architecture
Instead of trying to keep the HTTP and XMPP goroutines running
forever, let them crash when something goes wrong.  The main goroutine
notices the death and creates a new goroutine to take its place.

This architecture is easy to understand. It should be very resilient.
It also handles the case where a goroutine crashed because its
internal state got corrupted somehow.  Instead of trying to remedy
that corruption, we wipe the slate clean and make a new replacement.

I still don't feel right about the mutex for messages sent to the XMPP
goroutine.  It breaks the encapsulation of the XMPP goroutine.

See #12