Send message delivery receipts
When Twilio informs us that an outgoing SMS has been delivered to its
final destination, release the corresponding delivery receipt (if
there is one).  In XMPP clients that support message delivery
receipts, this usually triggers a change in the UI indicating that a
message has reached its final destination.

Now that we actually send delivery receipts, we can advertise XEP-0184
support in disco info queries and in our documentation.

See #5