Separate Component from Gateway
These two types share a lot of state, but they're semantically
distinct.  The Component is the piece which stays alive forever and
monitors the health of all other processes.  It also retains state
which should persist even if processes crash.

The Gateway translates between XMPP and HTTP.  It can crash if
needed.  When Component creates the Gateway, it lends ownership of
various state to the Gateway as long as it's alive.  When the Gateway
dies, ownership reverts back to the Component.

Because of the cleaner separation and these ownership rules, we can be
certain that data structures like receiptFor are never accessed
concurrently.

See #12