Try to establish a mutual presence subscription
Conversations only uses nicknames (XEP-0172) for the display name if
there's a mutual presence subscription between the parties.  When an
entity asks to subscribe to our presence, reciprocate by asking to
subscribe to theirs.

This doesn't add any caller ID support directly but paves the way for
it to be implemented.

See #9