Temporarily stop sending "subscribe" presence
I'm trying to convince Conversations to show XEP-0172 nicknames as the
display name for a newly arrived SMS from a contact we've never seen
before.  That's the most important use case for caller ID.

After several variations, this commit does it successfully.  I wanted
to record this code so that I can go back to it during future
attempts.

Another interesting observation from this debugging session: with
presence subscriptions enabled, Conversations shows the display name
correctly for a newly arrived message until I add that contact to my
roster.  At that point, it uses the JID.  That was a little unexpected
but otherwise fine.

Anyway, this commit is not viable long term because it makes the SMS
users appear unavailable in most chat clients.  In some (like Adium)
the SMS user is entirely hidden.  We need presence subscriptions so
that SMS users are easily accessible in most clients.

See #9