Better autoreply to "unsubscribe"
RFC 6121 is contradictory about how we should process an incoming
"unsubscribe" stanza.  Section 3.3.3 says that they must be silently
ignored unless `subFrom == yes` and it mentions nothing about an
"unsubscribed" autoreply.  Section A.3.2 mentions the autoreply and
requires a response for `subFrom != no`.

I've decided to only send autoreplies for a contact with an active
subscription.  Without a subscription, we shouldn't be sending
anything to the remote contact.

See #18