Support for XEP-0054 vcard-temp
When an incoming SMS arrives, remember the caller ID name.  When an
XMPP client asks for a vCard, respond with the previous caller ID
name.

Obviously this only works after we've received an incoming
SMS. However, that's the most useful case for caller ID anyway.  When
an outgoing SMS initiates a chat, the XMPP user had to provide the
phone number and they probably already know who it belongs to.

This vCard works with Swift and Adium.  Neither of them present a
smooth UI for it, but the name is visible upon request (choosing "View
Profile" or "Get Info", respectively).

See #9