Simplify XEP-0172 user nickname support
I just reread XEP-0172 section 4.2  It says that a nickname should be
sent in a first message (which we do), but it's silent about sending
nicknames with subsequent messages.  Since the spec allows it and
doing so simplifies the code substantially, we're now going to send
user nicknames with every message.

In practice, this seems to help Conversations to display the nickname
more reliably.  Other clients that I've tested with ignore the
nickname so it doesn't matter to them one way or the other.

See #9