Allow SMS providers to implement their own PSTN process logic
The `pstnProcess` type is abolished, and the HTTP server logic is moved
into the Twilio provider.  The `SmsProvider` interface is extended with
a new method:

```golang
RunPstnProcess(chan<- rxSms) <-chan struct{}
```

`RunPstnProcess` is responsible for starting whatever type of PSTN
process is appropriate for the provider, which for Twilio is an HTTP
server but for other providers might be something else.  This makes it
possible to implement SMS providers which don't receive SMSes over HTTP
(for example, SIP/SIMPLE).

This also makes it possible to integrate sms-over-xmpp into an existing
HTTP server by defining a custom `SmsProvider` that interfaces with the
existing server instead of starting its own.

Another motivation for this change is that other SMS providers that
use HTTP might not fit into the framework provided by `pstnProcess`.
For example, they might not use HTTP authentication, or they might
require a different response than the Twilio-specific one that was
introduced to `pstnProcess` in 179cdaf489f4a9c6b76aed876e0b08831f4d5bb3.
It is more flexible for the SMS provider to have complete control over
HTTP processing.