Remove unnecessary phone number <-> JID conversions
Previously, when receiving an SMS, the sender's phone number would be
converted to a JID using config.PhoneToAddress and used as the from
address for the XMPP message.  Using PhoneToAddress is unnecessary
because the from address should always be phone number at the component
domain.

Similarly, using config.AddressToPhone is unnecessary to determine the
destination phone number when sending an SMS; it's always the local part
of the JID.

Removing these unnecessary conversions will allow us to simplify
AddressToPhone and PhoneToAddress and fix a security vulnerability,
in a followup commit.