Don't try to guess JID or phone number if there's no entry in users/phones
Previously, if there was no entry for a phone number, we'd assume the
JID should be phone number at the component domain.  If there was no
entry for a JID, we'd assume the phone number was the local part.

This logic was insecure - an attacker at evil.com could send a message to
your sms-over-xmpp component with a from address of <your-phone-number>@evil.com,
and sms-over-xmpp would send an SMS from your number!  Now, messages from unknown
Jabber users will be ignored.

This logic was necessary prior to removing the unnecessary conversions
in commit 1def3ce96579e566aa8287057c735d26587e8f19.  Now that the unnecessary
conversions have been removed, this logic is not needed.

Since this logic was never documented, and the documentation told people
to explicitly configure their JID / phone number mappings, this change
should hopefully not break anyone.

Closes: #27